+++ tests/neptune/new/internal/backends/test_hosted_neptune_backend.py
@@ -93,7 +93,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
         response_error.errorDescription = "error1"
         swagger_client.api.executeOperations().response().result = [response_error]
         swagger_client.api.executeOperations.reset_mock()
-        upload_mock.return_value = FileUploadError("file1", "error2")
+        upload_mock.return_value = [FileUploadError("file1", "error2")]
         some_text = "Some streamed text"
         some_binary = b"Some streamed binary"
 
@@ -166,6 +166,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                             attribute="some/other/file.txt",
                             source="other/file/path.txt",
                             ext="txt",
+                            multipart_config=backend._client_config.multipart_config,
                         ),
                         call(
                             swagger_client=backend.leaderboard_client,
@@ -173,6 +174,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                             attribute="some/files/some_file",
                             source="path_to_file",
                             ext="",
+                            multipart_config=backend._client_config.multipart_config,
                         ),
                         call(
                             swagger_client=backend.leaderboard_client,
@@ -180,6 +182,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                             attribute="some/files/some_text_stream",
                             source=some_text.encode("utf-8"),
                             ext="txt",
+                            multipart_config=backend._client_config.multipart_config,
                         ),
                         call(
                             swagger_client=backend.leaderboard_client,
@@ -187,6 +190,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                             attribute="some/files/some_binary_stream",
                             source=some_binary,
                             ext="bin",
+                            multipart_config=backend._client_config.multipart_config,
                         ),
                     ],
                     any_order=True,
@@ -250,6 +254,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                             attribute="some/path/1/var",
                             source="/path/to/file",
                             ext="",
+                            multipart_config=backend._client_config.multipart_config,
                         ),
                         call(
                             swagger_client=backend.leaderboard_client,
@@ -257,6 +262,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                             attribute="some/path/2/var",
                             source="/some.file/with.dots.txt",
                             ext="txt",
+                            multipart_config=backend._client_config.multipart_config,
                         ),
                         call(
                             swagger_client=backend.leaderboard_client,
@@ -264,6 +270,7 @@ class TestHostedNeptuneBackend(unittest.TestCase, BackendTestMixin):
                             attribute="some/path/3/var",
                             source="/path/to/some_image.jpeg",
                             ext="jpeg",
+                            multipart_config=backend._client_config.multipart_config,
                         ),
                     ],
                     any_order=True,