+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [UNRELEASED] neptune-client 0.14.2
 
+### Features
+- Use new file upload API ([#789](https://github.com/neptune-ai/neptune-client/pull/789))
+
 ## Fixes
 - Fixed backward compatibility when syncing old offline data ([#810](https://github.com/neptune-ai/neptune-client/pull/810))
 
@@ -8,7 +11,6 @@
 ## Fixes
 - Fixed legacy url in NVML information ([#795](https://github.com/neptune-ai/neptune-client/pull/795))
 - Make init_project accepting kwargs only ([#805](https://github.com/neptune-ai/neptune-client/pull/805))
-- Use new file upload API ([#789](https://github.com/neptune-ai/neptune-client/pull/789))
 
 ## neptune-client 0.14.0