+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Use new file upload API ([#789](https://github.com/neptune-ai/neptune-client/pull/789))
 
 ## Fixes
+- Fixed listing available workspaces when invalid name was provided ([#818](https://github.com/neptune-ai/neptune-client/pull/818))
 - Fixed backward compatibility when syncing old offline data ([#810](https://github.com/neptune-ai/neptune-client/pull/810))
 
 ## neptune-client 0.14.1